+++ assets/r/transform_clean_delete_selected.R
@@ -1,0 +1,8 @@
+# Delete a variable. 
+
+# Remove specific variables from the dataset.
+
+ds$SELECTED_VAR <- NULL
+
+glimpse(ds)
+summary(ds)