+++ assets/r/transform_clean_delete_vars_missing.R
@@ -1,0 +1,6 @@
+# Delete columns with any missing values.
+
+ds <- ds[ , !(names(ds) %in% MISSING_VARS)]
+
+glimpse(ds)
+summary(ds)