+++ assets/r/transform_impute_mean_numeric.R
@@ -5,7 +5,7 @@
 # License: GNU General Public License, Version 3 (the "License")
 # https://www.gnu.org/licenses/gpl-3.0.en.html
 #
-# Time-stamp: <Saturday 2024-08-03 14:16:36 +1000 Graham Williams>
+# Time-stamp: <Sunday 2024-08-11 19:28:24 +1000 Graham Williams>
 #
 # Licensed under the GNU General Public License, Version 3 (the "License");
 #
@@ -30,7 +30,7 @@ if (is.numeric(ds$SELECTED_VAR))
 {
   ds %<>%
     mutate(IMN_SELECTED_VAR = ifelse(is.na(SELECTED_VAR),
-                                     mean(wind_gust_dir, na.rm = TRUE),
+                                     mean(SELECTED_VAR, na.rm = TRUE),
                                      SELECTED_VAR))
 }