+++ lib/widgets/image_page.dart
@@ -1,6 +1,6 @@
 /// Helper widget to build the common image based pages.
 //
-// Time-stamp: <Friday 2024-08-09 05:29:34 +1000 Graham Williams>
+// Time-stamp: <Saturday 2024-08-10 15:02:42 +1000 Graham Williams>
 //
 /// Copyright (C) 2024, Togaware Pty Ltd
 ///
@@ -29,6 +29,7 @@ library;
 
 import 'dart:io';
 import 'dart:async';
+import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
@@ -38,7 +39,9 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'package:rattle/constants/sunken_box_decoration.dart';
+import 'package:rattle/constants/temp_dir.dart';
 import 'package:rattle/utils/select_file.dart';
+import 'package:rattle/utils/show_image_dialog.dart';
 import 'package:rattle/utils/word_wrap.dart';
 
 class ImagePage extends StatelessWidget {
@@ -119,62 +122,7 @@ class ImagePage extends StatelessWidget {
                           color: Colors.blue,
                         ),
                         onPressed: () {
-                          // TODO 20240809 gjw for zy MOVE INTO SEPARATE FUNCTION/CLASS.
-                          //
-                          // By moving into a separate function/class we reduce
-                          // the cognitive overload of viewing the
-                          // logic/structure of this outer widget. I want to be
-                          // able to see the stucture at a glance on one screen,
-                          // not get lost in the detail.
-                          //
-                          showGeneralDialog(
-                            context: context,
-                            pageBuilder:
-                                (context, animation, secondaryAnimation) {
-                              return Center(
-                                child: Material(
-                                  type: MaterialType.transparency,
-                                  child: Stack(
-                                    children: [
-                                      Container(
-                                        width:
-                                            MediaQuery.of(context).size.width *
-                                                0.9,
-                                        height:
-                                            MediaQuery.of(context).size.height *
-                                                0.9,
-                                        padding: const EdgeInsets.all(16.0),
-                                        color: Colors.white,
-                                        child: InteractiveViewer(
-                                          maxScale: 5,
-                                          child: SvgPicture.memory(bytes),
-                                        ),
-                                      ),
-                                      Positioned(
-                                        top: 10,
-                                        right: 10,
-                                        child: GestureDetector(
-                                          onTap: () {
-                                            Navigator.of(context).pop();
-                                          },
-                                          child: const Icon(
-                                            Icons.close,
-                                            color: Colors.grey,
-                                            size: 30,
-                                          ),
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              );
-                            },
-                            barrierDismissible: true,
-                            barrierLabel: MaterialLocalizations.of(context)
-                                .modalBarrierDismissLabel,
-                            transitionDuration:
-                                const Duration(milliseconds: 200),
-                          );
+                          showImageDialog(context, bytes);
                         },
                         tooltip: 'Press here to view the plot\n'
                             'enlarged.',
@@ -186,7 +134,27 @@ class ImagePage extends StatelessWidget {
                         ),
                         tooltip: 'TODO Press here to view the plot\n'
                             'in a separate window.',
-                        onPressed: () {},
+                        onPressed: () {
+                          // Generate a unique file name for the new file in the
+                          // temporary directory.
+
+                          String fileName =
+                              'plot_${Random().nextInt(10000)}.svg';
+                          File tempFile = File('$tempDir/$fileName');
+
+                          // Copy the original file to the temporary file.
+
+                          File originalFile = File(path);
+                          tempFile
+                              .writeAsBytesSync(originalFile.readAsBytesSync());
+
+                          // Pop out a window to display the plot separate
+                          // to the Rattle app.
+
+                          Platform.isWindows
+                              ? Process.run('start', [tempFile.path])
+                              : Process.run('open', [tempFile.path]);
+                        },
                       ),
                       IconButton(
                         icon: const Icon(