+++ CHANGELOG.md
@@ -8,6 +8,17 @@ toward a 6.3 release.
 
 ## 6.3 FUTURE Wrangle tab implementation
 
++ Review and code cleanup [6.2.51 20240813 zy]
++ Extra string cleaning for Windows 10 [6.2.50 20240812 gjw]
++ Doc updates. Modernise R code. Trial await CONSOLE startup [6.2.49 20240812 gjw]
++ Bug fixes [6.2.48 20240811 gjw]
++ Re-engineer and rewrite CLEANUP [6.2.47 20240811 gjw]
++ Fix Windows version display. [6.2.46 20240810 bz]
++ Add CLEANUP. [6.2.46 20240810 yyx]
++ Implement image popout [6.2.45 20240810 gjw]
++ Move main.R conditional for Windows only [6.2.44 20240810 gjw] 
++ Move main.R into dataset_prep.R to fix Windows loading [6.2.43 20240809 gjw]
++ Fix Windows R console connectivity issue [6.2.41 20240809 Lutra-Fs]
 + Rearrange RESCALE interface [6.2.40 20240808 gjw]
 + Review and clean [6.2.39 20240808 zy]
 + Update configs for RESCALE and TREE [6.2.38 20240808 gjw]